Add 'last_modified' property to the ZipStream
The ZipStream now tracks the modified dates of added files and returns
the most recent one when accessing the `last_modified' property.
Currently non-files are assumed to have been modified at the time they
are added.

This `last_modified` property is used in the `zipserver` module to
provide a `Last-Modified` HTTP header in the response. This allows
clients to make a `HEAD` request to the server to check if anything in
the zip has changed without the server having to actually generate it.