Fix adding empty directories
Directly added empty directories (ie. not discovered via recursing) were
being added with an arcname of `directory/./`.

This was causing a mismatch in calculated vs. actual size. The size
calculation was using the length of `directory/./`, but the folder was
actually being added to the zip as just `directory` because of the
normalization that `zipfile.ZipInfo.from_file` performs on arcnames.

The fix is to normalize the arcnames in the same way that
`zipfile.ZipInfo.from_file` does before adding them to the zip. By doing
this, the size should always be correct, regardless of any issues with
creating arcnames. In this case, it also fixes the mangled directory
name.

Fixes #8