Allow specifying the size of iterables for sized ZipStreams
This enables an iterable to be added to a sized `ZipStream` without
immediately pulling it entirely into memory by providing the size of it
to `add()`. After that file entry is generated, the actual size of the
data from the iterable will be verified against the provided size and a
`RuntimeError` will be raised if they don't match.

This will allow for things like adding data from an HTTP stream with a
`Content-Length` header without having to fully download the file first
or give up on the `ZipStream` length calculation.

Closes #12