Fix get_info returning incorrect compression levels
When the compression level is not used (like for uncompressed or LZMA
compressed files), `get_info` would report whatever level was specified.
This has been resolved so it will now correctly report `None` for
compression types that don't support compression levels.

Also, as a small optimization, adding files without specifying a
compression type/level or specifying the same ones as the archive will
no longer have the type/level stored in the queue for processing.