Gracefully handle null bytes in arcnames
This fixes an issue where adding data with a null byte in its arcname
would cause the zipfile size to be incorrectly calculated. This happened
because the arcname would be trimmed to the first null byte when
generating the zipfile, but the size calculation would use the untrimmed
version.

This commit pre-trims and validates the arcname before it's added to the
list of files to add. This pre-validation also now provides a better
error for the case where a zero-length arcname is provided to
`ZipStream.mkdir` (previously `IndexError`, now the same as all other
invalid arcname messages)