Add the ability to query added and non-streamed files
This adds the `ZipStream.info_list` function which will return a list of
information on all the files that have been added to the stream. This
also includes information on files that have not been streamed yet.

This was added because `ZipStream.get_info` only returned information on
files that had already been streamed. The new `ZipStream.info_list`
function gives the ability to query the stream for things like number of
files, total size, etc. before starting to generate it.

In order to preserve compatibility `ZipStream.get_info` was not removed,
but will now emit a `DeprecationWarning` when used.