Prevent mismatches between arcname and data when adding paths (#22)
Previously, it was possible to add a directory and give it an arcname that
didn't end with a "/". Likewise, it was also possible to add a file
with a arcname that ended with a "/". These cases would cause an
incorrect stream size to be calculated, as well as create strange zip
files.

This commit fixes this issue by modifying arcnames as paths are added to
ensure that directories end with "/" and files do not.