Allow adding top-level files by using an arcname of "/"
Previously it was possible add the contents of a directory at the top
level of a zip by using an `arcname` of `"."`, but it wasn't documented
or discoverable. This commit adds explicit support for specifying that a
directory should have an `arcname` of `"/"`, therefore putting all of
its contents at the top level.

For example, to add every file in `/path/to/files/` to the top level of
a zip:
```
ZipStream.from_path("/path/to/files", arcname="/")
```

This also fixes an issue where leading path separators were not being
stripped in the same way as the `zipfile` module, potentially leading to
size mismatches.